refactor(api-gateway): filename corrected
